[Upd] Refactor stories for readability
